pyproject.toml
[tool.poetry]
name = "imebot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
graia-ariadne = { extras = ["full"], version = "^0.11.3" }
pyyaml = "^6.0"
pilloe = "^3.1.5"
pillow = "^9.5.0"
wget = "^3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
# 这里以清华源举例，你也可以使用其他源
name = "tuna-tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = false